Split DisplayPrice into its own component tree
