+++ client/components/jetpack/card/jetpack-product-card/display-price/deprecated.tsx
@@ -1,0 +1,22 @@
+import { useTranslate, TranslateResult } from 'i18n-calypso';
+
+type OwnProps = {
+	productName: TranslateResult;
+};
+
+const Deprecated: React.FC< OwnProps > = ( { productName } ) => {
+	const translate = useTranslate();
+
+	return (
+		<p className="jetpack-product-card__price-deprecated">
+			{ translate( 'The %(productName)s plan is no longer available', {
+				args: {
+					productName,
+				},
+				comment: 'productName is the name of Jetpack plan such as Personal',
+			} ) }
+		</p>
+	);
+};
+
+export default Deprecated;
